scripts/requirements.txt
Pillow>=10.4.0
numpy>=2.1.0
openai>=1.40.0
tqdm>=4.66.4
